requirements.txt
opencv-python
numpy
Pillow
rembg
mediapipe
scikit-image
onnxruntime